+++ _assets/sass/custom/_page.scss
@@ -37,6 +37,7 @@
     .usa-input {
       border: solid 1px;
       float: none;
+      height: 2.45rem;
     }
   }
