+++ _includes/search-box.html
@@ -1,8 +1,7 @@
-<div class="search-box-wrapper">
-    <label class="usa-label" for="searchbox">Search</label>
+<div class="search-box-wrapper margin-top-2">
     <div class="display-flex flex-row flex-align-center">
-        <input class="usa-input" type="search" id="searchbox" >
-        <button class="usa-button radius-right-lg" id="searchbtn">
+        <input class="usa-input margin-right-1 radius-lg" type="search" id="searchbox" >
+        <button class="usa-button radius-lg" id="searchbtn">
             Search
         </button>
     </div>
